extract src files into teacherNavigation/lessonMaterials/
