+++ apps/src/templates/teacherNavigation/lessonMaterials/ResourceIcon.tsx
