+++ apps/src/templates/teacherNavigation/lessonMaterials/ResourceIconType.tsx
