+++ apps/src/templates/teacherNavigation/lessonMaterials/lesson-materials.module.scss
